[None][refactor] move KV cache manager V2
