+++ tensorrt_llm/_torch/attention_backend/interface.py
@@ -25,7 +25,8 @@ from tensorrt_llm.models.modeling_utils import QuantConfig
 from ..memory_buffer_utils import Buffers
 from ..metadata import KVCacheParams
 from ..pyexecutor.mamba_cache_manager import BaseMambaCacheManager
-from ..pyexecutor.resource_manager import KVCacheManager, KVCacheManagerV2
+from ..pyexecutor.kv_cache_manager_v2 import KVCacheManagerV2
+from ..pyexecutor.resource_manager import KVCacheManager
 from ..utils import get_model_extra_attrs
 
 try:
