+++ tensorrt_llm/_torch/disaggregation/resource/cache_reuse.py
@@ -20,7 +20,8 @@ from typing import List, Sequence, Union
 import numpy as np
 
 from tensorrt_llm._torch.pyexecutor.llm_request import LlmRequest
-from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager, KVCacheManagerV2
+from tensorrt_llm._torch.pyexecutor.kv_cache_manager_v2 import KVCacheManagerV2
+from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
 
 from .page import AttentionLayerGroup
 from .utils import get_global_layer_ids
