+++ tensorrt_llm/_torch/pyexecutor/resource_manager.py
@@ -103,227 +103,6 @@ class ResourceManagerType(enum.Enum):
     SPEC_RESOURCE_MANAGER = "SPEC_RESOURCE_MANAGER"
 
 
-class Role:
-    KEY = DataRole("key")
-    VALUE = DataRole("value")
-    KEY_BLOCK_SCALE = DataRole("key_block_scale")
-    VALUE_BLOCK_SCALE = DataRole("value_block_scale")
-    ALL = DataRole("all")
-
-
-class _MmRunMetadata(NamedTuple):
-    """CPU tensors for exact multimodal run lookup.
-
-    `num_runs` is the total number of flat multimodal runs across all
-    multimodal items, i.e. `item_run_cu_offsets[-1]`. All members are 1-D CPU
-    int64 tensors with shape `[num_runs]`.
-    """
-
-    # Shape [num_runs]; full-prompt start index for each flat run.
-    run_positions: torch.Tensor
-    # Shape [num_runs]; one-past-last full-prompt index for each flat run.
-    run_ends: torch.Tensor
-    # Shape [num_runs]; logical multimodal item index for each flat run.
-    run_item_indices: torch.Tensor
-    # Shape [num_runs]; item-local token offset where each flat run begins.
-    run_item_offsets: torch.Tensor
-
-
-def _hash_to_digest(hash_ints: Sequence[int]) -> bytes:
-    # Convert 8 x int32 hash chunks to the 32-byte digest used by C++ block
-    # keys. The byte order matches getNthByte(), which extracts MSB first.
-    try:
-        hash_len = len(hash_ints)
-    except TypeError as exc:
-        raise ValueError(
-            "Expected 8 int32 hash values, got non-sized input") from exc
-    if hash_len != 8:
-        raise ValueError(f"Expected 8 int32 hash values, got {hash_len}")
-    if not all(isinstance(value, int) for value in hash_ints):
-        raise ValueError("Expected multimodal hash values to be integers")
-    return b''.join(v.to_bytes(4, 'big', signed=True) for v in hash_ints)
-
-
-def _ensure_int64_cpu_tensor(
-        values: Sequence[int] | torch.Tensor) -> torch.Tensor:
-    # Block-reuse augmentation is Python-side index math. The metadata is
-    # produced by host mm preprocessing and carried in
-    # MultimodalInput metadata. A non-CPU tensor here means the upstream
-    # contract drifted and would introduce an unexpected sync in the
-    # block-reuse path, so fail loudly.
-    if isinstance(values, torch.Tensor):
-        if values.device.type != "cpu":
-            raise ValueError(
-                "multimodal block-reuse metadata must be CPU-resident, "
-                f"got {values.device}")
-        return values.to(dtype=torch.int64)
-    return torch.as_tensor(values, dtype=torch.int64)
-
-
-def _resolve_multimodal_run_metadata(
-        req: LlmRequest) -> Optional[_MmRunMetadata]:
-    # TODO(perf): cache per request; block-reuse invokes this once per block,
-    # repeatedly rebuilding identical tensors for the same request metadata.
-    # Worked example for one logical multimodal item split by text:
-    #
-    #   prompt index: 0    1      2      3    4      5
-    #   prompt token: T0   MM0:0  MM0:1  T1   MM0:2  MM0:3
-    #   flat run:          run0   run0        run1   run1
-    #
-    # Inputs encode the prompt layout as:
-    #
-    #   item_run_cu_offsets = [0, 2]  # item 0 owns flat runs [0, 2)
-    #   run_positions       = [1, 4]  # run0 starts at prompt 1, run1 at 4
-    #   run_lengths         = [2, 2]  # token counts for run0 and run1
-    #
-    # The derived arrays below let the augmentation loop map any overlapping
-    # flat run back to the item digest and to the item-local token offset:
-    #
-    #   item_run_counts        = [2]     # item 0 has two runs
-    #   cumulative_run_lengths = [0,2,4] # tokens before run0, run1, and end
-    #   item_starts            = [0]     # item 0 starts at flat run 0
-    #   run_item_indices       = [0,0]   # run0 and run1 both belong to item 0
-    #   run_item_offsets       = [0,2]   # run1 begins at MM0 token offset 2
-    #
-    # So a chunk slice covering prompt [4, 6) uses MM0's digest with token
-    # offset 2; a slice covering prompt [5, 6) uses token offset 3.
-
-    # Shape [num_items + 1]; item i owns flat runs [offsets[i], offsets[i + 1]).
-    item_run_cu_offsets = req.multimodal_item_run_cu_offsets
-    # Shape [num_runs]; full-prompt start index for each flat run.
-    run_positions = req.multimodal_run_positions
-    # Shape [num_runs]; token count for each flat run.
-    run_lengths = req.multimodal_run_lengths
-
-    if all(field is None
-           for field in (item_run_cu_offsets, run_positions, run_lengths)):
-        return None
-
-    if (item_run_cu_offsets is None or run_positions is None
-            or run_lengths is None):
-        raise ValueError(
-            "multimodal run metadata must be validated before block reuse and provided together"
-        )
-
-    item_run_cu_offsets = _ensure_int64_cpu_tensor(item_run_cu_offsets)
-    run_positions = _ensure_int64_cpu_tensor(run_positions)
-    run_lengths = _ensure_int64_cpu_tensor(run_lengths)
-
-    # Shape [num_items]; number of flat runs owned by each logical item.
-    item_run_counts = item_run_cu_offsets[1:] - item_run_cu_offsets[:-1]
-
-    # Shape [num_runs + 1]; item-token count before each flat run, plus end.
-    cumulative_run_lengths = torch.cat(
-        (torch.zeros(1, dtype=torch.int64), torch.cumsum(run_lengths, dim=0)))
-
-    # Shape [num_items]; first flat-run index for each logical item.
-    item_starts = item_run_cu_offsets[:-1]
-
-    # Shape [num_runs]; logical multimodal item index for each flat run.
-    run_item_indices = torch.repeat_interleave(
-        torch.arange(item_run_counts.numel(), dtype=torch.int64),
-        item_run_counts)
-
-    # Shape [num_runs]; item-local token offset where each flat run begins.
-    run_item_offsets = (cumulative_run_lengths[:-1] -
-                        cumulative_run_lengths[item_starts][run_item_indices])
-    # Shape [num_runs]; one-past-last full-prompt index for each flat run.
-    run_ends = run_positions + run_lengths
-
-    return _MmRunMetadata(
-        run_positions=run_positions,
-        run_ends=run_ends,
-        run_item_indices=run_item_indices,
-        run_item_offsets=run_item_offsets,
-    )
-
-
-def _augment_tokens_with_mm_run_metadata(
-        vocab_size: int, result: list[TokenIdExt],
-        multimodal_hashes: Sequence[Sequence[int]], metadata: _MmRunMetadata,
-        chunk_start: int, chunk_end: int) -> list[TokenIdExt]:
-    # Only rewrite multimodal runs that overlap the materialized prompt slice.
-    overlap_mask = ((metadata.run_ends > chunk_start)
-                    & (metadata.run_positions < chunk_end))
-    # Shape [num_overlapping_runs]; flat-run indices that touch this chunk.
-    overlap_run_indices = torch.nonzero(overlap_mask).flatten()
-    if overlap_run_indices.numel() == 0:
-        return result
-
-    # `result` is indexed relative to this chunk, but multimodal cache-key
-    # tokens are indexed relative to the logical multimodal item. The rewrite
-    # below therefore computes, for each selected run segment:
-    # - chunk_result_offset = prompt position - chunk_start
-    # - item_token_offset = item-local run start + intra-run prompt offset
-    # Shape [num_overlapping_runs]; item index for each selected flat run.
-    overlap_run_item_indices = metadata.run_item_indices[overlap_run_indices]
-    # Materialize the selected run metadata once before the Python loop. These
-    # are CPU tensors by construction, and the loop below is request hot-path
-    # bookkeeping.
-    overlap_run_positions = metadata.run_positions[overlap_run_indices]
-    prompt_overlap_starts = torch.clamp(overlap_run_positions, min=chunk_start)
-    prompt_overlap_ends = torch.clamp(metadata.run_ends[overlap_run_indices],
-                                      max=chunk_end)
-    item_token_offsets = (metadata.run_item_offsets[overlap_run_indices] +
-                          prompt_overlap_starts - overlap_run_positions)
-    chunk_result_offsets = prompt_overlap_starts - chunk_start
-    lengths = prompt_overlap_ends - prompt_overlap_starts
-
-    current_item_idx: Optional[int] = None
-    digest = b""
-    for item_idx, chunk_result_offset, item_token_offset, length in zip(
-            overlap_run_item_indices.tolist(),
-            chunk_result_offsets.tolist(),
-            item_token_offsets.tolist(),
-            lengths.tolist(),
-            strict=True):
-        if item_idx != current_item_idx:
-            current_item_idx = item_idx
-            digest = _hash_to_digest(multimodal_hashes[item_idx])
-        # Feed the coarse item property (content digest) and granular run
-        # properties (item-local offset and span length) into the key
-        # generator, so cache keys reflect the actual multimodal tokens being
-        # rewritten.
-        result[chunk_result_offset:chunk_result_offset +
-               length] = gen_multimodal_cache_key_tokens(
-                   vocab_size, digest, length, token_offset=item_token_offset)
-
-    return result
-
-
-def _augment_tokens_with_contiguous_mm_metadata(
-        vocab_size: int, result: list[TokenIdExt],
-        multimodal_hashes: Sequence[Sequence[int]],
-        multimodal_positions: Sequence[int] | torch.Tensor,
-        multimodal_lengths: Sequence[int] | torch.Tensor, chunk_start: int,
-        chunk_end: int) -> list[TokenIdExt]:
-    # Legacy metadata assumes every item is one contiguous prompt span. Exact
-    # run metadata is preferred for video layouts that interleave text tokens.
-    positions = _ensure_int64_cpu_tensor(multimodal_positions)
-    lengths = _ensure_int64_cpu_tensor(multimodal_lengths)
-
-    item_ends = positions + lengths
-    overlap_mask = (item_ends > chunk_start) & (positions < chunk_end)
-    overlap_item_indices = torch.nonzero(overlap_mask).flatten()
-    for item_idx_tensor in overlap_item_indices:
-        item_idx = int(item_idx_tensor)
-        pos = positions[item_idx].item()
-        length = lengths[item_idx].item()
-        overlap_start = max(pos, chunk_start)
-        overlap_end = min(pos + length, chunk_end)
-        source_offset = overlap_start - pos
-        result_offset = overlap_start - chunk_start
-        overlap_length = overlap_end - overlap_start
-        result[result_offset:result_offset +
-               overlap_length] = gen_multimodal_cache_key_tokens(
-                   vocab_size,
-                   _hash_to_digest(multimodal_hashes[item_idx]),
-                   overlap_length,
-                   token_offset=source_offset)
-
-    return result
-
-
 def compute_page_count(token_count: int, tokens_per_page: int) -> int:
     return (token_count + tokens_per_page) // tokens_per_page
 
@@ -423,34 +202,6 @@ def request_context(is_draft: bool, scheduled_requests: ScheduledRequests):
     return RequestContext(is_draft, scheduled_requests)
 
 
-def _locate_accepted_draft_tokens(requests: List[LlmRequest]):
-    num_accepted_draft_tokens = []
-    accepted_draft_tokens_indices = []
-    rewind_draft_token_separate_adjustments = []
-    # for context requests, the py_num_accepted_draft_tokens = 0, and py_num_accepted_draft_tokens_indices = []
-    for seq in requests:
-        num_accepted_draft_tokens.append(seq.py_num_accepted_draft_tokens)
-        rewind_draft_token_separate_adjustments.append(
-            seq.py_rewind_draft_token_separate_adjustment)
-        accepted_draft_tokens_indices.extend(
-            seq.py_num_accepted_draft_tokens_indices)
-    batch_size = len(requests)
-    num_accepted_draft_tokens_offset = torch.zeros(batch_size + 1,
-                                                   dtype=torch.int32,
-                                                   device='cuda')
-    num_accepted_draft_tokens_offset[1:] = torch.cumsum(torch.tensor(
-        num_accepted_draft_tokens, dtype=torch.int32),
-                                                        dim=0)
-    accepted_draft_tokens_indices = torch.tensor(accepted_draft_tokens_indices,
-                                                 dtype=torch.int32,
-                                                 device='cuda')
-    rewind_draft_token_separate_adjustments = torch.tensor(
-        rewind_draft_token_separate_adjustments,
-        dtype=torch.int32,
-        device='cuda')
-    return num_accepted_draft_tokens_offset, accepted_draft_tokens_indices, rewind_draft_token_separate_adjustments
-
-
 # M-RoPE (Qwen2-VL/Qwen3-VL) splits positions across 3 axes: temporal/height/width.
 _MROPE_NUM_AXES = 3
 
@@ -482,57 +233,6 @@ def _populate_dummy_mrope_config(req: LlmRequest, token_num: int,
     if req.py_multimodal_data is None:
         req.py_multimodal_data = {}
     req.py_multimodal_data["mrope_config"] = mrope_config
-
-
-def _update_kv_cache_draft_token_location(cache_manager,
-                                          scheduled_batch: ScheduledRequests,
-                                          attn_metadata: "AttentionMetadata",
-                                          kv_cache_dtype_byte_size: float):
-    run_kv_cache_relocation = False
-    for request in scheduled_batch.generation_requests:
-        if request.state != LlmRequestState.GENERATION_COMPLETE:
-            if request.py_num_accepted_draft_tokens > 0 and len(
-                    request.py_num_accepted_draft_tokens_indices) > 0:
-                run_kv_cache_relocation = True
-    if not run_kv_cache_relocation:
-        return
-    requests = scheduled_batch.all_requests()
-    accepted_draft_token_offsets, packed_accepted_draft_tokens_indices, rewind_draft_token_separate_adjustments = _locate_accepted_draft_tokens(
-        requests)
-    past_key_value_lengths = attn_metadata.kv_lens_cuda[:len(requests)]
-    if attn_metadata.kv_cache_block_offsets is not None and attn_metadata.host_kv_cache_pool_pointers is not None and attn_metadata.host_kv_cache_pool_mapping is not None:
-        use_paged_kv_cache = True
-    else:
-        use_paged_kv_cache = False
-    assert use_paged_kv_cache, "Only paged kv cache is supported"
-    assert len(
-        cache_manager.max_attention_window_vec
-    ) == 1, "Currently, only one max attention window size is supported."
-
-    if use_paged_kv_cache:
-        assert len(set(cache_manager.num_kv_heads_per_layer)) == 1, \
-            "update_kv_cache_draft_token_location requires uniform num_kv_heads across all layers, " \
-            f"but got {cache_manager.num_kv_heads_per_layer}"
-        torch.ops.tensorrt_llm.update_kv_cache_draft_token_location(
-            accepted_draft_token_offsets,
-            packed_accepted_draft_tokens_indices,
-            past_key_value_lengths,
-            True,
-            cache_manager.num_layers,
-            # Use TP-sharded num_kv_heads (per-rank) instead of the unsharded
-            # total so the C++ kernel computes correct strides and grid dims.
-            cache_manager.num_kv_heads_per_layer[0],
-            int(cache_manager.head_dim * kv_cache_dtype_byte_size),
-            cache_manager.max_total_draft_tokens,
-            cache_manager.max_attention_window_vec[0],
-            rewind_draft_token_separate_adjustments,
-            None,
-            cache_manager.kv_cache_pool_pointers,
-            attn_metadata.kv_cache_block_offsets,
-            cache_manager.max_blocks_per_seq,
-            cache_manager.tokens_per_block,
-            None,
-        )
 
 
 class KVCacheManager(BaseResourceManager):
@@ -2327,1383 +2027,6 @@ class KVCacheManager(BaseResourceManager):
     def reset_reuse_state(self):
         """Reset the reuse state of the KV cache manager."""
         self.impl.reset_reuse_state()
-
-
-class KVCacheManagerV2(BaseResourceManager):
-
-    def __init__(
-        self,
-        kv_cache_config: KvCacheConfig,
-        kv_cache_type: CacheTypeCpp,
-        *,
-        num_layers: int,
-        num_kv_heads: Union[int, List[Optional[int]]],
-        head_dim: Union[int, List[int]],
-        tokens_per_block: int,
-        # Note that max_seq_len is not necessarily equal to kv_cache_config.num_tokens.
-        # It's derived from the model's BuildConfig for consistency with the C++ backend.
-        max_seq_len: int,
-        max_batch_size: int,
-        mapping: Mapping,
-        dtype: DataType = DataType.HALF,
-        spec_config=None,
-        layer_mask: Optional[List[bool]] = None,
-        vocab_size: int = None,
-        max_num_tokens: int = 8192,
-        model_config: Optional[ModelConfigCpp] = None,
-        max_beam_width: int = 1,
-        is_draft: bool = False,
-        kv_connector_manager: Optional[KvCacheConnectorManager] = None,
-        execution_stream: Optional[torch.cuda.Stream] = None,
-        is_disagg: bool = False,
-        **kwargs,
-    ) -> None:
-        self.mapping = mapping
-        self.dtype = dtype
-        self.is_disagg = is_disagg
-
-        assert kv_connector_manager is None, "kv_connector_manager is not supported for KVCacheManagerV2"
-        assert max_beam_width == 1, "max_beam_width must be 1 for KVCacheManagerV2"
-        assert not (mapping.cp_config.get('cp_type') == CpType.STAR), \
-            "Star attention is not supported for KVCacheManagerV2"
-
-        self.kv_cache_type = kv_cache_type
-        self.pp_layers, self.num_layers = get_pp_layers(
-            num_layers,
-            mapping,
-            spec_config=spec_config,
-            layer_mask=layer_mask,
-        )
-        self.is_draft = is_draft
-        self.num_local_layers = len(self.pp_layers)
-        self.layer_offsets = {
-            idx: offset
-            for offset, idx in enumerate(self.pp_layers)
-        }
-        self.max_beam_width = max_beam_width
-
-        tp_size = mapping.tp_size
-        if mapping.enable_attention_dp:
-            tp_size = 1
-
-        self.num_kv_heads = num_kv_heads
-        self.head_dim = head_dim
-        self.tokens_per_block = tokens_per_block
-        self.max_seq_len = max_seq_len
-        self.max_batch_size = max_batch_size
-        self.kv_factor = 1 if kv_cache_type == CacheTypeCpp.SELFKONLY else 2
-        from ..speculative import get_num_extra_kv_tokens
-        self.num_extra_kv_tokens = get_num_extra_kv_tokens(spec_config)
-        self.max_total_draft_tokens = spec_config.max_total_draft_tokens if spec_config is not None else 0
-
-        # Mirror V1's KV reserve sizing (see V1 __init__ for rationale).
-        self._kv_reserve_draft_tokens = self.max_total_draft_tokens
-        if (self.is_draft and spec_config is not None
-                and getattr(spec_config, 'use_dynamic_tree', False)
-                and getattr(spec_config, 'dynamic_tree_max_topK', 0) > 0):
-            draft_loop_tokens = spec_config.dynamic_tree_max_topK * spec_config.max_draft_len
-            self._kv_reserve_draft_tokens = max(self.max_total_draft_tokens,
-                                                draft_loop_tokens)
-
-        self.event_buffer_max_size = kv_cache_config.event_buffer_max_size
-
-        assert self.event_buffer_max_size == 0, "event_buffer_max_size must be 0"
-
-        self._stream = execution_stream if execution_stream is not None else torch.cuda.current_stream(
-        )
-        logger.info(f"[KVCacheManager] execution_stream: {self._stream}")
-
-        # Determine max_attention_window_vec
-        if kv_cache_config.max_attention_window is not None:
-
-            self.max_attention_window_vec = kv_cache_config.max_attention_window.copy(
-            )  # Make a copy to avoid modifying original
-            # Clamp all window sizes to max_seq_len before calculating the
-            # number of KV cache blocks. This prevents the KV cache pool from
-            # being skewed by the largest window values.
-            self.max_attention_window_vec = [
-                min(max_seq_len, w) for w in self.max_attention_window_vec
-            ]
-
-            self.max_attention_window_vec = [
-                None if w == max_seq_len else w
-                for w in self.max_attention_window_vec
-            ]
-
-        else:
-            self.max_attention_window_vec = [None]
-
-        if isinstance(num_kv_heads, int):
-            self.num_kv_heads_per_layer = [
-                (num_kv_heads + tp_size - 1) // tp_size
-                for _ in range(self.num_local_layers)
-            ]
-            self.total_num_kv_heads_per_layer = [
-                (num_kv_heads + tp_size - 1) // tp_size
-                for _ in range(self.num_layers)
-            ]
-        else:
-            assert len(num_kv_heads) == self.num_layers
-
-            def append_to_kv_heads_per_layer(num_kv_heads_per_layer: List[int],
-                                             kv_head: Optional[int]):
-                if kv_head is not None:
-                    num_kv_heads_per_layer.append(
-                        (kv_head + tp_size - 1) // tp_size)
-                else:
-                    num_kv_heads_per_layer.append(0)
-
-            self.num_kv_heads_per_layer = []
-            if self.num_local_layers > 0:
-                for i in self.pp_layers:
-                    kv_head = num_kv_heads[i]
-                    append_to_kv_heads_per_layer(self.num_kv_heads_per_layer,
-                                                 kv_head)
-
-            self.total_num_kv_heads_per_layer = []
-            for i in range(self.num_layers):
-                kv_head = num_kv_heads[i]
-                append_to_kv_heads_per_layer(self.total_num_kv_heads_per_layer,
-                                             kv_head)
-
-        # Build per-layer head_dim (similar to num_kv_heads_per_layer)
-        if isinstance(head_dim, int):
-            self.head_dim_per_layer = [
-                head_dim for _ in range(self.num_local_layers)
-            ]
-        else:
-            assert len(head_dim) == self.num_layers, \
-                f"head_dim list length ({len(head_dim)}) must match num_layers ({self.num_layers})"
-            self.head_dim_per_layer = []
-            if self.num_local_layers > 0:
-                for i in self.pp_layers:
-                    self.head_dim_per_layer.append(head_dim[i])
-            if len(set(self.head_dim_per_layer)) > 1:
-                logger.info(
-                    f"Per-layer head_dim: {len(self.head_dim_per_layer)} layers, "
-                    f"unique values={set(self.head_dim_per_layer)}")
-
-        self.is_vswa = len(set(self.max_attention_window_vec)) > 1
-
-        quota = float('inf')
-        if kv_cache_config.max_gpu_total_bytes is not None and kv_cache_config.max_gpu_total_bytes > 0:
-            quota = int(kv_cache_config.max_gpu_total_bytes)
-            logger.info(
-                f"max_gpu_total_bytes is provided. New quota is {quota / (1 << 30)}GiB"
-            )
-        if kv_cache_config.max_tokens is not None:
-            quota_from_max_tokens = int(
-                math.ceil(
-                    self._get_quota_from_max_tokens(kv_cache_config.max_tokens)
-                    / kv_cache_config.max_util_for_resume))
-            quota = min(quota, quota_from_max_tokens)
-            logger.info(
-                f"max_tokens {kv_cache_config.max_tokens} is provided. Allowed quota from max_tokens is {quota_from_max_tokens / (1 << 30)}GiB. New quota is {quota / (1 << 30)}GiB"
-            )
-
-        assert quota != float(
-            'inf'
-        ), "Quota not set. Check kv_cache_config.max_tokens or kv_cache_config.max_gpu_total_bytes"
-
-        # Sync KV cache token capacity across ranks so all ranks allocate
-        # the same number of tokens and the scheduler produces identical
-        # batches.  Normalize to token count before the allreduce because
-        # bytes_per_token varies across PP ranks (different local layers).
-        if mapping.world_size > 1:
-            dist = Distributed.get(mapping)
-            bytes_per_token = self.get_cache_bytes_per_token()
-            max_tokens = quota / bytes_per_token
-            max_tokens = dist.allreduce(max_tokens, op=ReduceOp.MIN)
-            quota = max_tokens * bytes_per_token
-
-        logger.info(
-            f"KV cache manager v2 device quota set to {quota / (1 << 30)}GiB")
-
-        cache_tiers: List[CacheTierConfig] = [GpuCacheTierConfig(quota=quota)]
-        if kv_cache_config.host_cache_size is not None and kv_cache_config.host_cache_size > 0:
-            host_quota = kv_cache_config.host_cache_size
-        else:
-            # The V2 MAX_UTILIZATION scheduler relies on suspend/resume to
-            # evict and later restore KV cache pages.  Without a host tier,
-            # suspended pages have nowhere to be offloaded and resume()
-            # always fails, causing a scheduling deadlock where no
-            # generation request can ever make progress.
-            #
-            # Automatically provision a host tier matching the GPU quota so
-            # suspend/resume works out of the box.  Cap at available host
-            # memory to avoid allocation failures.
-            try:
-                mem_available = os.sysconf('SC_PAGE_SIZE') * os.sysconf(
-                    'SC_AVPHYS_PAGES')
-            except (ValueError, OSError):
-                mem_available = float('inf')
-            host_quota = min(quota, int(mem_available * 0.5))
-            if host_quota <= 0:
-                host_quota = quota
-        if host_quota > 0:
-            cache_tiers.append(HostCacheTierConfig(quota=host_quota))
-            logger.info(
-                f"KV cache manager v2 host cache quota set to {host_quota / (1 << 30):.2f}GiB"
-            )
-        disk_cache_size = kv_cache_config.disk_cache_size
-        if disk_cache_size is not None and disk_cache_size > 0:
-            disk_cache_path = kv_cache_config.disk_cache_path
-            assert disk_cache_path is not None
-            cache_tiers.append(
-                DiskCacheTierConfig(quota=disk_cache_size,
-                                    path=disk_cache_path))
-            logger.info(
-                f"KV cache manager v2 disk cache quota set to {disk_cache_size / (1 << 30):.2f}GiB at {disk_cache_path}"
-            )
-
-        self.vocab_size = vocab_size
-
-        config = self._build_cache_config(
-            kv_cache_config,
-            tokens_per_block=tokens_per_block,
-            vocab_size=vocab_size,
-            cache_tiers=cache_tiers,
-        )
-
-        self.kv_cache_manager_py_config = config
-
-        self.impl = KVCacheManagerPy(config)
-
-        self.num_pools = len(self.impl.layer_grouping)
-
-        num_layers = len(config.layers)
-        self.layer_to_pool_mapping_dict: dict[int, int] = {
-            layer_id: self.impl.get_layer_group_id(layer_id)
-            for layer_id in typed_range(LayerId(num_layers))
-        }
-
-        (self.kv_cache_pool_pointers,
-         self.kv_cache_pool_mapping) = self._build_pool_mapping_tensors()
-
-        self.kv_cache_map: dict[int, _KVCache] = {}
-
-        # Tracks the draft length allocated by try_allocate_generation per
-        # request.  Used by extend_capacity_for_tokens to compute the exact
-        # padding delta instead of blindly extending, which would cause
-        # unbounded capacity growth.
-        self._allocated_draft_lens: dict[int, int] = {}
-
-        # Defensive cap for get_num_available_tokens: when host cache is
-        # enabled, clamp_max_seq_len_for_mem may return a value that spans
-        # both GPU and host tiers.  Storing the explicit max_tokens (if set)
-        # lets us cap the result to GPU-only capacity so callers like CUDA
-        # graph warmup don't over-allocate beyond the GPU pool.
-        # None when max_tokens is not explicitly configured — other config
-        # paths (max_gpu_total_bytes, free_gpu_memory_fraction) are already
-        # bounded by the GPU quota passed to GpuCacheTierConfig.
-        self._gpu_max_tokens = kv_cache_config.max_tokens
-
-        max_num_tokens = self.get_num_available_tokens(
-            token_num_upper_bound=max_seq_len)
-
-        if max_seq_len > max_num_tokens:
-            logger.warning(
-                f"max_seq_len {max_seq_len} is greater than max_num_tokens {max_num_tokens} that can be allocated in kv cache manager, setting max_seq_len to {max_num_tokens}"
-            )
-            # max_num_tokens is a float from clamp_max_seq_len_for_mem; cast
-            # so downstream int-only consumers (torch.randint size, range)
-            # stay int.
-            self.max_seq_len = int(max_num_tokens)
-
-        # Pad max_blocks_per_seq to next multiple of 4 (copy_block_offsets kernel).
-        # Account for max single-sequence capacity = seq_len + extra KV tokens +
-        # _kv_reserve_draft_tokens (see __init__) + 1 base decode token.
-        max_seq_capacity = self.max_seq_len + self.num_extra_kv_tokens + self._kv_reserve_draft_tokens + 1
-        self.max_blocks_per_seq = (max_seq_capacity + tokens_per_block -
-                                   1) // tokens_per_block
-        if self.max_blocks_per_seq % 4 != 0:
-            self.max_blocks_per_seq = ((self.max_blocks_per_seq + 3) // 4) * 4
-
-        self.enable_block_reuse = kv_cache_config.enable_block_reuse
-        self.enable_partial_reuse = kv_cache_config.enable_partial_reuse
-
-        # With pipeline parallelism, multiple microbatches can be in-flight
-        # simultaneously, so we need slots for all concurrent sequences.
-        # Plus 1 for cuda graph dummy request.
-        # In disaggregated mode, use a coefficient of 2: at any moment up to
-        # `max_num_sequences` requests can be actively generating while another
-        # up to `max_num_sequences` requests are still in KV transfer
-        # (TRANS_IN_PROGRESS) and continue to hold their index slots. The 2x
-        # capacity lets the next batch of active requests acquire slots without
-        # waiting for the previous batch's transfers to finish.
-        max_num_sequences = max_batch_size * mapping.pp_size
-        index_mapper_capacity = max_num_sequences * (2 if is_disagg else 1) + 1
-        logger.info(
-            f"KVCacheManagerV2: IndexMapper capacity={index_mapper_capacity} "
-            f"(max_num_sequences={max_num_sequences}, is_disagg={is_disagg}, max_beam_width={max_beam_width})"
-        )
-        self.index_mapper = IndexMapper(index_mapper_capacity, max_beam_width)
-        self._early_freed_index_requests: set[int] = set()
-        self.index_scales = torch.empty(self.num_pools,
-                                        dtype=torch.int32,
-                                        pin_memory=prefer_pinned(),
-                                        device='cpu')
-        self.kv_offset = torch.empty(self.num_pools,
-                                     dtype=torch.int32,
-                                     pin_memory=prefer_pinned(),
-                                     device='cpu')
-        for pool_id in range(self.num_pools):
-            layer_id = self.impl.layer_grouping[pool_id][0]
-            self.index_scales[pool_id] = self.impl.get_page_index_scale(
-                layer_id, Role.KEY)
-            if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-                self.kv_offset[pool_id] = exact_div(
-                    self.impl.get_mem_pool_base_address(layer_id, Role.VALUE) -
-                    self.impl.get_mem_pool_base_address(layer_id, Role.KEY),
-                    self.impl.get_page_stride(layer_id, Role.KEY))
-            else:
-                self.kv_offset[pool_id] = 0
-
-        # Keep unused block offsets as safe block index 0.
-        self.host_kv_cache_block_offsets = torch.zeros(
-            self.num_pools,
-            index_mapper_capacity * max_beam_width,
-            2,  # key and value
-            self.max_blocks_per_seq,
-            dtype=torch.int32,
-            pin_memory=prefer_pinned(),
-            device='cpu')
-
-    def _get_quota_from_max_tokens(self, max_tokens: int) -> int:
-        return int(max_tokens * self.get_cache_bytes_per_token())
-
-    def _build_pool_mapping_tensors(self) -> Tuple[torch.Tensor, torch.Tensor]:
-        kv_cache_pool_pointers = torch.tensor([[
-            self.impl.get_mem_pool_base_address(
-                self.impl.layer_grouping[pool_id][0], Role.KEY), 0
-        ] for pool_id in range(self.num_pools)],
-                                              dtype=torch.int64,
-                                              device="cpu",
-                                              pin_memory=prefer_pinned())
-
-        if self.dtype == DataType.NVFP4:
-            kv_cache_pool_pointers = torch.stack([
-                kv_cache_pool_pointers,
-                torch.tensor([[
-                    self.impl.get_mem_pool_base_address(
-                        self.impl.layer_grouping[pool_id][0],
-                        Role.KEY_BLOCK_SCALE), 0
-                ] for pool_id in range(self.num_pools)],
-                             dtype=torch.int64,
-                             device="cpu",
-                             pin_memory=prefer_pinned())
-            ],
-                                                 dim=-1)
-
-        kv_cache_pool_mapping_list = []
-        for layer_id in typed_range(LayerId(self.num_local_layers)):
-            layer_group_id = self.impl.get_layer_group_id(layer_id)
-            if self.dtype != DataType.NVFP4:
-                addr_offset = self.impl.get_mem_pool_base_address(
-                    layer_id, Role.KEY) - int(
-                        kv_cache_pool_pointers[layer_group_id][0])
-            else:
-                addr_offset = self.impl.get_mem_pool_base_address(
-                    layer_id, Role.KEY) - int(
-                        kv_cache_pool_pointers[layer_group_id][0][0])
-                block_scale_addr_offset = self.impl.get_mem_pool_base_address(
-                    layer_id, Role.KEY_BLOCK_SCALE) - int(
-                        kv_cache_pool_pointers[layer_group_id][0][1])
-                block_scale_offset = exact_div(
-                    block_scale_addr_offset,
-                    self.get_layer_bytes_per_token(
-                        layer_id, Role.KEY_BLOCK_SCALE) * self.kv_factor *
-                    self.tokens_per_block)
-            offset = exact_div(
-                addr_offset,
-                self.get_layer_bytes_per_token(layer_id, Role.KEY) *
-                self.kv_factor * self.tokens_per_block)
-
-            if self.dtype == DataType.NVFP4:
-                assert block_scale_offset == offset, "Block scale offset and offset should be the same"
-
-            kv_cache_pool_mapping_list.append([layer_group_id, offset])
-
-        kv_cache_pool_mapping = torch.tensor(kv_cache_pool_mapping_list,
-                                             dtype=torch.int32,
-                                             device="cpu",
-                                             pin_memory=prefer_pinned())
-        return kv_cache_pool_pointers, kv_cache_pool_mapping
-
-    def _build_cache_config(
-        self,
-        kv_cache_config: KvCacheConfig,
-        *,
-        tokens_per_block: int,
-        vocab_size: Optional[int],
-        cache_tiers: List[CacheTierConfig],
-    ) -> KVCacheManagerConfigPy:
-        buffer_type = [Role.KEY]
-        if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-            buffer_type.append(Role.VALUE)
-        if kv_cache_config.dtype == "nvfp4":
-            for layer_idx, hd in enumerate(self.head_dim_per_layer):
-                assert hd % 2 == 0, \
-                    f"head_dim must be divisible by 2 for nvfp4 kv cache, but layer {layer_idx} has head_dim={hd}"
-            buffer_type.append(Role.KEY_BLOCK_SCALE)
-            if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-                buffer_type.append(Role.VALUE_BLOCK_SCALE)
-
-        return KVCacheManagerConfigPy(
-            tokens_per_block=tokens_per_block,
-            vocab_size=vocab_size,
-            cache_tiers=cache_tiers,
-            max_util_for_resume=kv_cache_config.max_util_for_resume,
-            layers=[
-                AttentionLayerConfig(
-                    layer_id=layer_id,
-                    buffers=[
-                        BufferConfig(
-                            role=role,
-                            size=self.get_layer_bytes_per_token(
-                                local_layer_idx=layer_id, data_role=role) *
-                            tokens_per_block,
-                        ) for role in buffer_type
-                    ],
-                    sliding_window_size=self.max_attention_window_vec[
-                        self.pp_layers[layer_id] %
-                        len(self.max_attention_window_vec)],
-                    num_sink_tokens=None,
-                ) for layer_id in typed_range(LayerId(self.num_local_layers))
-            ],
-        )
-
-    @property
-    def blocks_in_primary_pool(self) -> int:
-        """
-        Get the number of blocks in the primary pool.
-        """
-        return self.impl.get_page_index_upper_bound(0, Role.KEY)
-
-    def get_buffers(self,
-                    layer_idx: int,
-                    kv_layout: str = "NHD") -> Optional[torch.Tensor]:
-        layer_offset = self.layer_offsets[layer_idx]
-        addr_key = self.impl.get_mem_pool_base_address(layer_offset, Role.KEY)
-        if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-            addr_value = self.impl.get_mem_pool_base_address(
-                layer_offset, Role.VALUE)
-            page_size_key = self.impl.get_page_stride(layer_offset, Role.KEY)
-            page_size_value = self.impl.get_page_stride(layer_offset,
-                                                        Role.VALUE)
-
-            assert addr_key + page_size_value == addr_value and page_size_key == page_size_value
-
-        assert kv_layout in ["NHD",
-                             "HND"], f"Unsupported kv_layout: {kv_layout}"
-
-        element_per_container = 1
-        dtype = self.dtype
-        if dtype == DataType.NVFP4:
-            element_per_container = 2
-            dtype = torch.int8
-
-        layer_head_dim = self.head_dim_per_layer[layer_offset]
-        if kv_layout == "NHD":
-            shape = [
-                self.impl.get_page_index_upper_bound(layer_offset, Role.KEY) //
-                self.kv_factor,
-                self.kv_factor,
-                self.tokens_per_block,
-                self.num_kv_heads_per_layer[layer_offset],
-                layer_head_dim // element_per_container,
-            ]
-        else:
-            shape = [
-                self.impl.get_page_index_upper_bound(layer_offset, Role.KEY) //
-                self.kv_factor,
-                self.kv_factor,
-                self.num_kv_heads_per_layer[layer_offset],
-                self.tokens_per_block,
-                layer_head_dim // element_per_container,
-            ]
-
-        return convert_to_torch_tensor(TensorWrapper(
-            addr_key,
-            dtype,
-            shape,
-        ))
-
-    def get_num_available_tokens(self,
-                                 *,
-                                 token_num_upper_bound: int,
-                                 batch_size: int = 1,
-                                 max_num_draft_tokens: int = 0) -> int:
-        extra_tokens = self.num_extra_kv_tokens + max_num_draft_tokens
-        # Token num upper bound is the maximum number of tokens that can be allocated in the kv cache manager.
-        # We need to add extra tokens to the token num upper bound to account for the extra tokens.
-        clamped = self.impl.clamp_max_seq_len_for_mem(
-            batch_size, token_num_upper_bound + extra_tokens) - extra_tokens
-        # clamp_max_seq_len_for_mem considers all tiers (GPU + host).  When
-        # max_tokens is explicitly set, cap by GPU-only capacity so callers
-        # (e.g. CUDA graph warmup) don't exceed the GPU pool.
-        if self._gpu_max_tokens is not None:
-            clamped = min(clamped, self._gpu_max_tokens - extra_tokens)
-        return clamped
-
-    def get_num_free_blocks(self) -> int:
-        # NOTE This method is used to get the number of blocks in the primary pool not the FREE blocks.
-        # However, since we only use this function when the kv cache manager is empty, so it is safe to do so.
-        assert len(
-            self.kv_cache_map
-        ) == 0, "get_num_free_blocks is only used when the kv cache manager is empty"
-        max_num_pages = max([
-            self.impl.get_page_index_upper_bound(layer_id, Role.KEY)
-            for layer_id in typed_range(LayerId(self.num_local_layers))
-        ])
-        return max_num_pages // self.kv_factor
-
-    # ---- Scheduling API (called by KVCacheV2Scheduler) ----
-
-    def is_request_active(self, request_id: int) -> bool:
-        """Return True if *request_id* has a live, non-suspended KV cache."""
-        kv_cache = self.kv_cache_map.get(request_id)
-        return kv_cache is not None and kv_cache.is_active
-
-    def _required_gen_capacity(self, req: LlmRequest,
-                               current_capacity: int) -> int:
-        """Compute generation KV cache capacity for a request.
-
-        Grows *current_capacity* by 1 + draft tokens.
-        """
-        draft_len = get_draft_token_length(req)
-        return current_capacity + 1 + draft_len
-
-    def try_allocate_generation(self, req: LlmRequest) -> bool:
-        """Try to allocate one additional KV cache slot for a generation request.
-
-        Resumes from suspended state if needed, then resizes capacity by 1 (+
-        draft tokens). Returns True on success, False if allocation failed.
-        """
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is None:
-            return False
-
-        if not kv_cache.is_active:
-            if not kv_cache.resume(self._stream.cuda_stream):
-                return False
-            self._restore_page_index_bufs(req.py_request_id, kv_cache)
-
-        draft_len = get_draft_token_length(req)
-        self._allocated_draft_lens[req.py_request_id] = draft_len
-        return kv_cache.resize(
-            self._required_gen_capacity(req, kv_cache.capacity))
-
-    def revert_allocate_generation(self, req: LlmRequest) -> None:
-        """Undo the capacity growth from try_allocate_generation.
-
-        When attention DP causes can_queue=False after scheduling, the
-        forward pass is skipped but the scheduler already grew each
-        generation request's KV cache capacity by 1 (+draft tokens).
-        This method shrinks capacity back to undo that spurious growth
-        so it does not accumulate across iterations and overflow the
-        host page-index buffer.
-        """
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is None or not kv_cache.is_active:
-            return
-        draft_len = get_draft_token_length(req)
-        reverted_cap = kv_cache.capacity - 1 - draft_len
-        if reverted_cap < 0:
-            return
-        if not kv_cache.resize(reverted_cap):
-            raise RuntimeError(
-                f"Failed to revert KV cache capacity for request "
-                f"{req.py_request_id} from {kv_cache.capacity} to "
-                f"{reverted_cap}")
-
-    def revert_allocate_context(self, req: LlmRequest) -> None:
-        """Undo the capacity growth from this iter's ``resize_context``.
-
-        When delay batching (``_balance_adp_requests`` /
-        ``_waiting_requests``) defers a context request after V2
-        scheduling, the forward pass is skipped for that request but the
-        scheduler already grew its KV cache capacity to cover the chunk.
-        This shrinks capacity back to the pre-resize value so the
-        freshly-allocated pages can be reused during the wait window —
-        important for long contexts where one deferred request can hold
-        GBs of KV.
-        """
-        pre_cap = getattr(req, "py_ctx_pre_resize_cap", None)
-        if pre_cap is None:
-            return
-        # Mark as consumed even if the resize below is skipped, so a
-        # later iter does not see a stale snapshot.
-        req.py_ctx_pre_resize_cap = None
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is None or not kv_cache.is_active:
-            return
-        if pre_cap >= kv_cache.capacity:
-            return
-        if not kv_cache.resize(pre_cap):
-            raise RuntimeError(
-                f"Failed to revert KV cache capacity for context "
-                f"request {req.py_request_id} from "
-                f"{kv_cache.capacity} to {pre_cap}")
-        if pre_cap > 0:
-            kv_cache.suspend()
-
-    def _restore_page_index_bufs(self, request_id: int, kv_cache) -> None:
-        """Re-connect host page-index buffers after resume().
-
-        suspend() clears the base_page_index_buf pointers (sets them to
-        None) so the KV cache stops writing page indices to the host
-        buffer.  After resume(), the KV cache has re-locked pages but
-        copy_batch_block_offsets still reads from the host buffer, so we
-        must re-connect the buffers to avoid stale/zero page indices that
-        would cause illegal memory accesses during the forward pass.
-        """
-        index = self.index_mapper.get_index(request_id)
-        for i in range(self.max_beam_width):
-            for pool_idx in range(self.num_pools):
-                buffer: torch.Tensor = self.host_kv_cache_block_offsets[
-                    pool_idx, index * self.max_beam_width + i, 0]
-                kv_cache.set_base_page_index_buf(i, pool_idx,
-                                                 memoryview(buffer.numpy()))
-
-    def _resume_and_restore(self, req_id: int, kv_cache) -> bool:
-        """Resume a suspended KV cache and restore its page index buffers.
-
-        Returns True if the cache is (or becomes) active, False on failure.
-        """
-        if kv_cache.is_active:
-            return True
-        if not kv_cache.resume(self._stream.cuda_stream):
-            return False
-        self._restore_page_index_bufs(req_id, kv_cache)
-        return True
-
-    def prepare_context(self, req: LlmRequest) -> bool:
-        """Create _KVCache, handle block reuse, and resume. Does NOT resize.
-
-        For first chunk: creates _KVCache (with block reuse lookup if enabled),
-        sets context_current_position, and resumes from suspended state.
-        For subsequent chunks: verifies existing cache is active.
-        Returns True on success, False if preparation failed.
-        """
-        if req.is_first_context_chunk:
-            kv_cache = self.kv_cache_map.get(req.py_request_id)
-            if kv_cache is None:
-                # Last token cannot be recovered, so we don't include it in
-                # the input tokens to look up for the block that can be reused.
-                if self.enable_block_reuse:
-                    all_tokens = req.get_tokens(DEFAULT_BEAM_INDEX)
-                    tokens = self._augment_tokens_for_block_reuse(
-                        all_tokens, req, end=len(all_tokens) - 1)
-                else:
-                    tokens = None
-                kv_cache = self._create_kv_cache(req.py_request_id,
-                                                 req.lora_task_id,
-                                                 tokens,
-                                                 cache_salt=req.cache_salt)
-                if kv_cache is None:
-                    return False
-                kv_cache.cuda_stream = self._stream.cuda_stream
-
-            if not self.enable_block_reuse:
-                kv_cache.stop_committing()
-            else:
-                req.context_current_position = kv_cache.num_committed_tokens
-                req.set_prepopulated_prompt_len(kv_cache.num_committed_tokens,
-                                                self.tokens_per_block)
-
-            return self._resume_and_restore(req.py_request_id, kv_cache)
-        else:
-            # Subsequent chunk: cache must exist from first chunk.
-            # It may be suspended (e.g., evicted between chunks), so
-            # _resume_and_restore handles reactivation.
-            kv_cache = self.kv_cache_map.get(req.py_request_id)
-            assert kv_cache is not None, (
-                f"KV cache missing for non-first context chunk, request {req.py_request_id}"
-            )
-            return self._resume_and_restore(req.py_request_id, kv_cache)
-
-    def resize_context(self, req: LlmRequest, num_tokens: int) -> bool:
-        """Resize KV cache to cover context_current_position + num_tokens.
-
-        num_tokens is the number of tokens to be processed (i.e.,
-        context_remaining_length or a chunk thereof). The target capacity is
-        computed as context_current_position + num_tokens so that block reuse
-        overlaps with existing capacity are handled correctly.
-        Returns True on success, False if resize failed (first chunk is
-        suspended on failure).
-
-        Snapshots the pre-resize capacity on ``req.py_ctx_pre_resize_cap``
-        when growth happens so ``revert_allocate_context`` can undo it if
-        delay batching defers the request.
-        """
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is None:
-            return False
-
-        target = req.context_current_position + num_tokens + self.num_extra_kv_tokens
-        capacity = max(kv_cache.capacity, target)
-        pre_cap = kv_cache.capacity
-
-        if not kv_cache.resize(capacity):
-            if req.is_first_context_chunk:
-                kv_cache.suspend()
-            return False
-
-        # None means "no growth this iter, nothing to revert"; this also
-        # invalidates a stale snapshot from a prior iter on the same req.
-        req.py_ctx_pre_resize_cap = pre_cap if capacity > pre_cap else None
-        return True
-
-    def extend_capacity_for_tokens(self, request: LlmRequest) -> None:
-        """Extend KV cache capacity for the CUDA-graph padding delta.
-
-        ``try_allocate_generation`` allocated capacity for the schedule-reduced
-        draft length.  After padding restores ``py_draft_tokens`` to the static
-        max, we must extend by exactly the difference so that the subsequent
-        rewind (which operates on the padded length) does not underflow.
-
-        The delta is computed from ``_allocated_draft_lens`` (recorded by
-        ``try_allocate_generation``) vs the current draft length (post-padding).
-        """
-        allocated = self._allocated_draft_lens.pop(request.py_request_id, None)
-        if allocated is None:
-            return
-        current_draft_len = get_draft_token_length(request)
-        delta = current_draft_len - allocated
-        if delta <= 0:
-            return
-        kv_cache = self.kv_cache_map[request.py_request_id]
-        new_capacity = kv_cache.capacity + delta
-        success = kv_cache.resize(new_capacity)
-        if not success:
-            raise ValueError(
-                f"Failed to extend capacity of KV cache for request "
-                f"{request.py_request_id} by {delta} tokens "
-                f"(target capacity {new_capacity})")
-
-    def suspend_request(self, req: LlmRequest) -> None:
-        """Suspend a request's KV cache (move to host tier)."""
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is not None and kv_cache.is_active:
-            kv_cache.suspend()
-
-    def resume_request(self, req: LlmRequest) -> bool:
-        """Resume a previously-suspended KV cache for *req*.
-
-        Returns True if the cache is (or becomes) active on GPU, False if
-        resume was refused (e.g. GPU pressure above max_util_for_resume)
-        or no cache exists for the request.
-        """
-        kv_cache = self.kv_cache_map.get(req.py_request_id)
-        if kv_cache is None:
-            return False
-        return self._resume_and_restore(req.py_request_id, kv_cache)
-
-    # ---- prepare_resources ----
-
-    @nvtx_range("prepare_resources_kv_cache_manager_v2")
-    def prepare_resources(self, scheduled_batch: ScheduledRequests):
-        if self.is_draft:
-            # Draft V2 manager: mirror the main manager by creating/resizing
-            # KV caches for scheduled requests (the main V2 scheduler does not
-            # know about the draft manager).
-            self._prepare_draft_resources(scheduled_batch)
-            return
-
-    def _prepare_draft_resources(self, scheduled_batch: ScheduledRequests):
-        """Create/resize KV caches in the draft V2 manager for scheduled requests.
-
-        The main V2 scheduler only manages the primary KV cache manager.
-        The draft manager must mirror context/generation allocations so that
-        its IndexMapper contains the correct request IDs for
-        copy_batch_block_offsets().
-        """
-        with request_context(True, scheduled_batch):
-            for req in scheduled_batch.context_requests:
-                kv_cache = self.kv_cache_map.get(req.py_request_id)
-                if kv_cache is None:
-                    kv_cache = self._create_kv_cache(req.py_request_id,
-                                                     req.lora_task_id,
-                                                     None,
-                                                     cache_salt=req.cache_salt)
-                    kv_cache.stop_committing()
-                if not self._resume_and_restore(req.py_request_id, kv_cache):
-                    raise RuntimeError(
-                        f"Failed to resume draft KV cache for request {req.py_request_id}"
-                    )
-                draft_len = get_draft_token_length(req)
-                capacity = (req.context_current_position +
-                            req.context_chunk_size + draft_len +
-                            self.num_extra_kv_tokens)
-                if not kv_cache.resize(capacity):
-                    raise RuntimeError(
-                        f"Draft KV cache context resize failed for request "
-                        f"{req.py_request_id}: could not resize to {capacity} tokens"
-                    )
-
-            for req in scheduled_batch.generation_requests:
-                kv_cache = self.kv_cache_map.get(req.py_request_id)
-                if kv_cache is None:
-                    raise RuntimeError(
-                        f"Missing draft KV cache for generation request {req.py_request_id}"
-                    )
-                if not self._resume_and_restore(req.py_request_id, kv_cache):
-                    raise RuntimeError(
-                        f"Failed to resume draft KV cache for request {req.py_request_id}"
-                    )
-                new_cap = self._required_gen_capacity(req, kv_cache.capacity)
-                # Pad the resize up to _kv_reserve_draft_tokens (see __init__);
-                # no-op when reserve == draft_token_length.
-                reserve_slack = (self._kv_reserve_draft_tokens -
-                                 get_draft_token_length(req))
-                if reserve_slack > 0:
-                    new_cap += reserve_slack
-                if not kv_cache.resize(new_cap):
-                    raise RuntimeError(
-                        f"Draft KV cache generation resize failed for request "
-                        f"{req.py_request_id}: could not resize to {new_cap} tokens"
-                    )
-
-    def _augment_tokens_for_block_reuse(
-            self,
-            tokens: Sequence[int],
-            req: LlmRequest,
-            start: int = 0,
-            end: int | None = None) -> Sequence[TokenIdExt]:
-        """Augment token sequence with multimodal content digests for block reuse.
-
-        Multimodal placeholder tokens (e.g. image_token_id) share the same ID
-        regardless of the underlying content. This method replaces each
-        multimodal token region with TokenIdExt values produced by
-        gen_multimodal_cache_key_tokens(), embedding the content digest
-        (Blake3 hash) into the token sequence so that the radix tree can
-        distinguish blocks belonging to different images/videos.
-
-        When *start*/*end* are given, they define the chunk bounds; only
-        `tokens[start:end]` is materialized and returned. This avoids
-        re-augmenting the full prompt on every chunk during chunked prefill.
-
-        For text-only requests this is a no-op.
-        """
-        if end is None:
-            end = len(tokens)
-        chunk_start = start
-        chunk_end = end
-        is_sliced = chunk_start != 0 or chunk_end != len(tokens)
-
-        if (req.multimodal_hashes is None or req.multimodal_positions is None
-                or req.multimodal_lengths is None):
-            return tokens[chunk_start:chunk_end] if is_sliced else tokens
-
-        result: list[TokenIdExt] = list(tokens[chunk_start:chunk_end])
-        run_metadata = _resolve_multimodal_run_metadata(req)
-        if run_metadata is not None:
-            return _augment_tokens_with_mm_run_metadata(self.vocab_size, result,
-                                                        req.multimodal_hashes,
-                                                        run_metadata,
-                                                        chunk_start, chunk_end)
-
-        return _augment_tokens_with_contiguous_mm_metadata(
-            self.vocab_size, result, req.multimodal_hashes,
-            req.multimodal_positions, req.multimodal_lengths, chunk_start,
-            chunk_end)
-
-    def get_kv_cache_stats(self):
-        kv_cache_stats = KvCacheStats()
-        kv_cache_stats.allocated_bytes = self.impl.get_quota(GPU_LEVEL)
-
-        return kv_cache_stats
-
-    def get_iteration_stats(self):
-        """V2 does not support per-iteration stats yet."""
-        return None
-
-    def get_block_ids_per_seq(self, request_ids: List[int]) -> torch.Tensor:
-        block_ids_per_seq = self.get_batch_cache_indices(request_ids)
-        block_ids_per_seq_tensors = [
-            torch.tensor([
-                i // self.num_local_layers if i != BAD_PAGE_INDEX else 0
-                for i in sublist
-            ],
-                         dtype=torch.int) for sublist in block_ids_per_seq
-        ]
-        padded_tensor = torch.nn.utils.rnn.pad_sequence(
-            block_ids_per_seq_tensors, batch_first=True, padding_value=0)
-        return padded_tensor
-
-    def add_dummy_requests(
-            self,
-            request_ids: List[int],
-            # Note that token_nums should be past_kv_len + input_len (without
-            # spec decoding). The draft tokens will be added in this function,
-            # so we don't need to take care of it in the caller. When preparing
-            # token_nums, we should not take the draft tokens into account, so
-            # don't use the kv_cache_manager.max_seq_len, which includes both
-            # extra tokens and draft tokens.
-            token_nums: Optional[List[int]] = None,
-            is_gen: bool = False,
-            prepare_resource: bool = True,
-            max_num_draft_tokens: int = 0,
-            kv_reserve_draft_tokens: Optional[int] = None,
-            use_mrope: bool = False,
-            max_beam_width: int = 1,
-            num_extra_decoding_steps: int = 0,
-            draft_kv_cache_manager: Optional['BaseResourceManager'] = None):
-        _kv_draft = kv_reserve_draft_tokens if kv_reserve_draft_tokens is not None else max_num_draft_tokens
-
-        beam_width = max_beam_width
-        requests = []
-
-        def release_resources(current_request: LlmRequest,
-                              free_draft_resources: bool = False) -> None:
-            for req in requests:
-                self.free_resources(req)
-            self.free_resources(current_request)
-            if draft_kv_cache_manager is not None:
-                for req in requests:
-                    draft_kv_cache_manager.free_resources(req)
-                if free_draft_resources:
-                    draft_kv_cache_manager.free_resources(current_request)
-
-        for i, req_id in enumerate(request_ids):
-            # exact choice of n can be ignored for dummy requests
-            sampling_params = SamplingParams(n=beam_width,
-                                             best_of=beam_width,
-                                             use_beam_search=beam_width > 1)
-            # Here 1+max_num_draft_tokens is used to extend the prompt length to
-            # a non-zero number to skip illegal memory access issue in MLA kernel
-            # during warmup.
-            token_num = token_nums[
-                i] if token_nums is not None else 1 + max_num_draft_tokens
-            # token_num - 1 is the past history length in generation.
-            history_hint = max(0, token_num - 1) if is_gen else None
-            # TODO: support cross attention
-            encoder_input_tokens = None
-            # Using 1 instead of 0 prevents NaN during warmup in e.g. Deepseek
-            input_tokens = [1 for _ in range(token_num)]
-            req = LlmRequest(request_id=req_id,
-                             max_new_tokens=1,
-                             input_tokens=input_tokens,
-                             sampling_config=SamplingConfig(
-                                 sampling_params._get_sampling_config()),
-                             is_streaming=False,
-                             encoder_input_tokens=encoder_input_tokens)
-            req.is_dummy_request = True
-            req.paged_kv_block_ids = []
-            if prepare_resource:
-                # Dummy/warmup request. ``stop_committing()`` below blocks all
-                # writes to the radix tree, so the choice of branch does not
-                # affect committed state. ``cache_salt`` is left defaulted
-                # to None to avoid coupling synthetic data to any salted branch.
-                kv_cache = self._create_kv_cache(req.py_request_id,
-                                                 req.lora_task_id, input_tokens)
-                assert kv_cache.num_committed_tokens == 0
-                success = kv_cache.resume(self._stream.cuda_stream)
-                if not success:
-                    release_resources(req)
-                    return None
-                kv_cache.stop_committing()
-                dummy_capacity = token_num + self.num_extra_kv_tokens + num_extra_decoding_steps
-                # Need to hint the committed history to activate stale-block
-                # optimization and match the solver's pool budget.
-                success = kv_cache.resize(dummy_capacity,
-                                          history_length=history_hint)
-                if not success:
-                    release_resources(req)
-                    return None
-                draft_kv_cache = None
-                if draft_kv_cache_manager is not None:
-                    draft_kv_cache = draft_kv_cache_manager._create_kv_cache(
-                        req.py_request_id, req.lora_task_id, input_tokens)
-                    # Dummy path: see comment above, no salt.
-                    success = draft_kv_cache.resume(
-                        draft_kv_cache_manager._stream.cuda_stream)
-                    if not success:
-                        release_resources(req, free_draft_resources=True)
-                        return None
-                    draft_kv_cache.stop_committing()
-                    success = draft_kv_cache.resize(dummy_capacity)
-                    if not success:
-                        release_resources(req, free_draft_resources=True)
-                        return None
-
-            if is_gen:
-                req.state = LlmRequestState.GENERATION_IN_PROGRESS
-                req.prompt_len = token_num - 1
-                req.py_prompt_len = req.prompt_len
-                req.py_draft_tokens = [1] * max_num_draft_tokens
-                if prepare_resource:
-                    new_capacity = kv_cache.capacity + _kv_draft + 1
-                    success = kv_cache.resize(new_capacity,
-                                              history_length=history_hint)
-                    if not success:
-                        release_resources(req,
-                                          free_draft_resources=draft_kv_cache
-                                          is not None)
-                        return None
-                    if draft_kv_cache is not None:
-                        success = draft_kv_cache.resize(new_capacity)
-                        if not success:
-                            release_resources(req, free_draft_resources=True)
-                            return None
-
-            if use_mrope:
-                _populate_dummy_mrope_config(req, token_num, is_gen)
-            requests.append(req)
-
-        return requests
-
-    def try_commit_blocks_for_reuse(self, request: LlmRequest,
-                                    kv_cache) -> None:
-        if (self.enable_block_reuse and not self.is_draft
-                and not request.is_dummy_request
-                and request.context_current_position
-                > kv_cache.num_committed_tokens):
-            tokens = self._augment_tokens_for_block_reuse(
-                request.get_tokens(DEFAULT_BEAM_INDEX),
-                request,
-                start=kv_cache.num_committed_tokens,
-                end=request.context_current_position)
-            kv_cache.commit(tokens)
-            kv_cache.stop_committing()
-
-    def release_index_slot(self, request_id: int) -> None:
-        """Release IndexMapper slot early while keeping KV cache blocks allocated.
-
-        After prefill completes on a context-only worker, the IndexMapper slot
-        (used for host_kv_cache_block_offsets during model forward) is no longer
-        needed.  Releasing it early allows new requests to be scheduled while
-        the KV cache blocks are still being transferred via NIXL/UCX.
-        """
-        self.index_mapper.remove_sequence(request_id)
-        self._early_freed_index_requests.add(request_id)
-
-    def free_resources(self, request: LlmRequest, pin_on_release: bool = False):
-        self._allocated_draft_lens.pop(request.py_request_id, None)
-        kv_cache = self.kv_cache_map.pop(request.py_request_id, None)
-        if kv_cache is None:
-            return
-        self.try_commit_blocks_for_reuse(request, kv_cache)
-        kv_cache.close()
-        if request.py_request_id in self._early_freed_index_requests:
-            self._early_freed_index_requests.discard(request.py_request_id)
-        else:
-            self.index_mapper.remove_sequence(request.py_request_id)
-
-    def get_batch_cache_indices(
-            self,
-            request_ids: List[int],
-            layer_idx: Optional[int] = None) -> List[List[int]]:
-        if layer_idx is None:
-            pool_id = 0
-        else:
-            pool_id = self.layer_to_pool_mapping_dict[
-                self.layer_offsets[layer_idx]]
-        return self._get_batch_cache_indices_by_pool_id(request_ids,
-                                                        pool_id=pool_id,
-                                                        is_kv_aggregate=True)
-
-    def _get_batch_cache_indices_by_pool_id(
-            self,
-            request_ids: List[int],
-            *,
-            pool_id: int = 0,
-            is_kv_aggregate: bool = True) -> List[List[int]]:
-
-        if is_kv_aggregate:
-            # Div by kv_factor to index kv cache with size [num_blocks, kv_factor, tokens_per_block, num_kv_heads, head_dim]
-            div_factor = self.kv_factor
-        else:
-            div_factor = 1
-
-        res = []
-
-        for req_id in request_ids:
-            idx_tensor = torch.as_tensor(
-                self.kv_cache_map[req_id].get_base_page_indices(pool_id))
-            res.append((torch.where(
-                idx_tensor != BAD_PAGE_INDEX,
-                idx_tensor * self.index_scales[pool_id] // div_factor,
-                BAD_PAGE_INDEX)).tolist())
-
-        return res
-
-    def get_cache_bytes_per_token(self) -> int:
-        data_roles = [Role.KEY]
-        if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-            data_roles.append(Role.VALUE)
-        if self.dtype == DataType.NVFP4:
-            data_roles.append(Role.KEY_BLOCK_SCALE)
-            if self.kv_cache_type != CacheTypeCpp.SELFKONLY:
-                data_roles.append(Role.VALUE_BLOCK_SCALE)
-
-        return sum(
-            self.get_layer_bytes_per_token(local_layer_idx=local_layer_idx,
-                                           data_role=data_role)
-            for local_layer_idx in range(self.num_local_layers)
-            for data_role in data_roles)
-
-    def get_layer_bytes_per_token(self, local_layer_idx: int, data_role: Role):
-        if self.dtype not in (
-                DataType.FP8,
-                DataType.HALF,
-                DataType.BF16,
-                DataType.FLOAT,
-                DataType.NVFP4,
-        ):
-            raise ValueError(f"Cannot support {self.dtype} KV cache.")
-
-        if data_role == Role.ALL:
-            kv_factor = self.kv_factor
-        elif data_role in [
-                Role.KEY, Role.VALUE, Role.KEY_BLOCK_SCALE,
-                Role.VALUE_BLOCK_SCALE
-        ]:
-            if data_role in [Role.KEY_BLOCK_SCALE, Role.VALUE_BLOCK_SCALE]:
-                assert self.dtype == DataType.NVFP4, "NVFP4 is the only supported dtype for block quant data roles"
-            if data_role == Role.VALUE:
-                assert self.kv_cache_type != CacheTypeCpp.SELFKONLY, "VALUE data role is not supported for SELFKONLY cache type"
-            kv_factor = 1
-        else:
-            raise ValueError(f"Invalid data role: {data_role}")
-
-        cache_size_per_token = kv_factor * self.num_kv_heads_per_layer[
-            local_layer_idx] * self.head_dim_per_layer[local_layer_idx]
-
-        cache_size_bytes_per_token = get_size_in_bytes(cache_size_per_token,
-                                                       self.dtype)
-
-        if data_role in [Role.KEY, Role.VALUE]:
-            return cache_size_bytes_per_token
-
-        quant_size_per_token = 0
-
-        if self.dtype == DataType.NVFP4:
-            quant_size_per_token = self.calculate_scaling_factor_size_bytes(
-                cache_size_per_token,
-                quant_vector_size=16,
-                scaling_factor_dtype=DataType.FP8,
-            )
-
-        if data_role in [Role.KEY_BLOCK_SCALE, Role.VALUE_BLOCK_SCALE]:
-            return quant_size_per_token
-
-        # Role.ALL combines both
-        return cache_size_bytes_per_token + quant_size_per_token
-
-    @staticmethod
-    def calculate_scaling_factor_size_bytes(
-            cache_size: int, quant_vector_size: int,
-            scaling_factor_dtype: DataType) -> int:
-        assert cache_size % quant_vector_size == 0, "NVFP4 cache size must be divisible by quant vector size"
-        return get_size_in_bytes(cache_size // quant_vector_size,
-                                 scaling_factor_dtype)
-
-    def check_invalid_values_in_kv_cache(self,
-                                         fill_with_zero: bool = False) -> bool:
-        some_checks_unavailable = False
-        has_invalid_values = torch.tensor([False],
-                                          dtype=torch.bool,
-                                          device=torch.cuda.current_device())
-        pool_handled = set()
-
-        # Handle each layer from start to end to traverse the whole KV cache.
-        for layer_id, layer_offset in self.layer_offsets.items():
-            pool_id = self.layer_to_pool_mapping_dict[layer_offset]
-            if pool_id in pool_handled:
-                continue
-            buffer = self.get_buffers(layer_id)
-            # process in chunks of 256 pages to avoid OoM
-            for i in range(0, buffer.shape[0], 256):
-                buffer_slice = buffer[i:i + 256]
-                try:
-                    has_invalid_values.logical_or_(
-                        torch.isnan(buffer_slice).any())
-                    has_invalid_values.logical_or_(
-                        torch.isinf(buffer_slice).any())
-                except NotImplementedError:
-                    some_checks_unavailable = True
-            if fill_with_zero:
-                buffer.zero_()
-            pool_handled.add(pool_id)
-        torch.cuda.synchronize()
-
-        if some_checks_unavailable:
-            logger.warning(
-                "`torch.isnan` or `torch.isinf` is not implemented for current kv cache dtype, related checks are skipped"
-            )
-        return bool(has_invalid_values)
-
-    def shutdown(self):
-        for kv_cache in self.kv_cache_map.values():
-            kv_cache.close()
-        self.kv_cache_map.clear()
-        self.impl.shutdown()
-
-    def get_max_resource_count(self) -> int:
-        # TODO: implement this
-        return 1
-
-    def get_needed_resource_to_completion(self, request: LlmRequest) -> int:
-        # TODO: implement this
-        # context_token_count = request.orig_prompt_len
-        # num_context_blocks = context_token_count // self.tokens_per_block
-        # remaining_tokens = context_token_count + request.max_new_tokens - num_context_blocks * self.tokens_per_block
-        # need_blocks = num_context_blocks + math.ceil(
-        #     remaining_tokens / self.tokens_per_block)
-        # return need_blocks
-        return 0
-
-    # TODO: refactor get_cache_size_per_token and get_cache_bytes_per_token to use the same logic
-    @staticmethod
-    def get_cache_size_per_token(model_config: ModelConfigPython,
-                                 mapping: Mapping,
-                                 num_layers: Optional[int] = None,
-                                 **kwargs):
-        # get kv cache dtype bytes
-        mem_per_token = 2
-        quant_config = model_config.quant_config
-        if quant_config is not None and quant_config.quant_mode.has_fp8_kv_cache(
-        ):
-            mem_per_token = 1
-
-        # get num key value heads
-        config = model_config.pretrained_config
-        num_key_value_heads = getattr(config, 'num_key_value_heads',
-                                      config.num_attention_heads)
-        if isinstance(num_key_value_heads, Iterable):
-            num_key_value_heads = sum(num_key_value_heads) / len(
-                num_key_value_heads)
-
-        # get head dim
-        mla = hasattr(config,
-                      "kv_lora_rank") and config.kv_lora_rank is not None
-        if mla:
-            head_dim = config.kv_lora_rank + config.qk_rope_head_dim
-            kv_factor = 1
-        else:
-            tp_size = 1 if mapping.enable_attention_dp else mapping.tp_size
-            head_dim = getattr(config, "head_dim", None)
-            if not isinstance(head_dim, int):
-                head_dim = config.hidden_size // config.num_attention_heads
-            head_dim = head_dim * num_key_value_heads // tp_size
-            kv_factor = 2
-
-        num_attention_layers = KVCacheManager._resolve_num_attention_layers(
-            model_config, mapping, num_layers)
-        mem_per_token *= num_attention_layers * head_dim
-
-        # K and V
-        mem_per_token *= kv_factor
-        return mem_per_token
-
-    def update_resources(self,
-                         scheduled_batch: ScheduledRequests,
-                         attn_metadata: "AttentionMetadata" = None,
-                         kv_cache_dtype_byte_size: float = None):
-        if not self.is_draft:
-            _update_kv_cache_draft_token_location(self, scheduled_batch,
-                                                  attn_metadata,
-                                                  kv_cache_dtype_byte_size)
-        for req in scheduled_batch.context_requests:
-            if req.py_request_id not in self.kv_cache_map:
-                continue
-            kv_cache = self.kv_cache_map[req.py_request_id]
-            # In the overlap scheduler, iteration N+1's eviction may
-            # suspend a ctx request's KV cache while iteration N's
-            # update_resources still needs to process it.  Skip the
-            # resize — the request will be resumed by the scheduler
-            # on the next iteration.
-            if not kv_cache.is_active:
-                continue
-            if self.enable_block_reuse and not self.is_draft and not req.is_dummy_request:
-                if req.context_current_position > kv_cache.num_committed_tokens:
-                    tokens = self._augment_tokens_for_block_reuse(
-                        req.get_tokens(DEFAULT_BEAM_INDEX),
-                        req,
-                        start=kv_cache.num_committed_tokens,
-                        end=req.context_current_position)
-                    kv_cache.commit(tokens)
-                if req.context_remaining_length == 0:
-                    kv_cache.stop_committing()
-            else:
-                success = kv_cache.resize(None, req.context_current_position)
-                if not success:
-                    raise ValueError(
-                        f"Failed to resize history length of KV cache for request {req.py_request_id} to {req.context_current_position} tokens at context update"
-                    )
-
-        for req in scheduled_batch.generation_requests:
-            if req.py_request_id not in self.kv_cache_map:
-                continue
-            kv_cache = self.kv_cache_map[req.py_request_id]
-            # In the overlap scheduler, the scheduler for iteration N+1
-            # may suspend a gen request's KV cache (via self-eviction or
-            # victim eviction) while iteration N's update_resources still
-            # needs to process it. Skip suspended caches — the request
-            # will be resumed by the scheduler on the next iteration.
-            if not kv_cache.is_active:
-                continue
-            new_capacity = None if req.state in (
-                LlmRequestState.GENERATION_COMPLETE,
-                LlmRequestState.CONTEXT_INIT
-            ) else kv_cache.capacity - req.py_rewind_len
-            success = kv_cache.resize(new_capacity, req.max_beam_num_tokens - 1)
-            if not success:
-                raise ValueError(
-                    f"Failed to resize KV cache for request {req.py_request_id} to capacity {new_capacity} and history length {req.max_beam_num_tokens - 1} tokens at generation update"
-                )
-
-    def copy_batch_block_offsets(self, dst_tensor: torch.Tensor,
-                                 request_ids: List[int], beam_width: int,
-                                 num_contexts: int, num_seqs: int):
-        assert beam_width == 1, "beam_width must be 1 for KVCacheManagerV2"
-
-        copy_idx = self.index_mapper.get_copy_index(request_ids, num_contexts,
-                                                    beam_width)
-        assert copy_idx.shape[0] == num_seqs
-
-        copy_batch_block_offsets_to_device(self.host_kv_cache_block_offsets,
-                                           dst_tensor, copy_idx,
-                                           self.index_scales, self.kv_offset,
-                                           self._stream.cuda_stream)
-
-    def _create_kv_cache(self,
-                         request_id: int,
-                         lora_task_id: int | None,
-                         input_tokens: Sequence[TokenIdExt] | None,
-                         cache_salt: str | None = None):
-        assert request_id not in self.kv_cache_map, f"KV cache for request {request_id} already exists"
-        if self.index_mapper.num_free_slots() == 0:
-            logger.warning(
-                "No free IndexMapper slots for request %s "
-                "(%d/%d slots in use, likely held by DISAGG_GENERATION_TRANS_IN_PROGRESS requests). "
-                "Skipping KV cache creation; request will retry next iteration.",
-                request_id, self.index_mapper.size(), self.index_mapper.size())
-            return None
-        # ReuseScope.salt is int|None; derive a deterministic int from the
-        # cache_salt string so the same string yields the same reuse namespace
-        # across processes (matches C++ blockKey hashing on cacheSalt).
-        salt_int = (int.from_bytes(
-            hashlib.sha256(cache_salt.encode("utf-8")).digest()[:8], "little")
-                    if cache_salt is not None else None)
-        kv_cache = self.impl.create_kv_cache(
-            ReuseScope(lora_id=lora_task_id, salt=salt_int),
-            input_tokens,
-        )
-        self.kv_cache_map[request_id] = kv_cache
-        index = self.index_mapper.add_new_sequence(request_id)
-        for i in range(self.max_beam_width):
-            for pool_idx in range(self.num_pools):
-                buffer: torch.Tensor = self.host_kv_cache_block_offsets[
-                    pool_idx, index * self.max_beam_width + i, 0]
-                kv_cache.set_base_page_index_buf(i, pool_idx,
-                                                 memoryview(buffer.numpy()))
-        return kv_cache
-
-    def reset_reuse_state(self):
-        self.impl.clear_reusable_blocks()
 
 
 class SlotManager:
