+++ tests/unittest/_torch/executor/test_kv_cache_estimation.py
@@ -12,7 +12,7 @@ from unittest.mock import Mock, patch
 import pytest
 
 from tensorrt_llm._torch.pyexecutor._util import CacheCost, KvCacheCreator
-from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManagerV2
+from tensorrt_llm._torch.pyexecutor.kv_cache_manager_v2 import KVCacheManagerV2
 
 # ---------------------------------------------------------------------------
 # Helpers
